Move timeseries fields to subfolder
