+++ bundles/admin/admin-layereditor/view/AdminLayerForm/VisualizationTabPane/TimeSeries/TimeSeriesMetadataToggleLevel.jsx
